refactor: 클래스명, 패키지명 변경
